# app/swagger.yaml
openapi: 3.0.0
info:
  title: My REST API
  version: 1.0.0
  description: A REST API to manage projects, tasks, teams, and users.
servers:
  - url: http://localhost:3000/
    description: Localhost development server
components:
  schemas:
...
    Team:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        username:
          type: string
        users:
          type: array
          items:
            type: string
        owner:
          type: string
paths:
...
  /team/create:
    post:
      summary: Create a new team
      operationId: createTeam
      tags:
        - Teams
      requestBody:
        required: true
        content:
          application/json: 
            schema:
              type: object
              required:
                - name
                - description
                - username
              properties:
                name:
                  type: string
                description:
                  type: string
                username:
                  type: string
      responses:
        201:
          description: The team was successfully created
        500:
          description: Creating the team encountered a problem
...
  /team/invite/{teamID}/{username}:
    get:
      summary: Invite a user to a team
      operationId: inviteUserToTeam
      tags:
        - Teams
      parameters:
        - in: path
          name: teamID
          type: string
          required: true
        - in: path
          name: username
          type: string
          required: true
      responses:
        200:
          description: The request was successfully registered
        400:
          description: The user you are trying to invite to the team was not found or has already been invited
  /team/{id}:
    get:
      summary: Get a specific team by ID
      operationId: getTeamById
      tags:
        - Teams
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        200:
          description: The requested team
          schema:
            $ref: '#/components/schemas/Team'
        404:
          description: No team was found
  /team/remove/{id}:
    delete:
      summary: Remove a team by ID
      operationId: removeTeamById
      tags:
        - Teams
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        200:
          description: The team was successfully deleted
        404:
          description: No team was found
        500:
          description: The team was not deleted. Please try again
...